refactor: update user model and migrations, remove unused fields, and adjust foreign key references
