+++ backend/accounts/models.py
@@ -1,13 +1,14 @@
-from django.contrib.auth.models import AbstractUser
 from django.db import models
 
-class User(AbstractUser):
+class User(models.Model):
     ROLE_CHOICES = [
         ('admin', 'Admin'),
         ('staff', 'Staff'),
     ]
     
     name = models.CharField(max_length=255, null=True, blank=True)
+    username = models.CharField(max_length=150, unique=True)
+    password = models.CharField(max_length=128)
     role = models.CharField(max_length=10, choices=ROLE_CHOICES, default='staff', null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
